+++ app_feup/lib/view/Widgets/bus_stop_card.dart
@@ -8,6 +8,7 @@ import 'package:uni/view/Pages/bus_stop_selection_page.dart';
 import 'package:uni/view/Widgets/bus_stop_row.dart';
 import 'package:uni/view/Widgets/last_update_timestamp.dart';
 import 'package:uni/view/Widgets/row_container.dart';
+import 'package:uni/utils/constants.dart' as Constants;
 
 import 'generic_card.dart';
 
@@ -18,10 +19,10 @@ class BusStopCard extends GenericCard {
       : super.fromEditingInformation(key, editingMode, onDelete);
 
   @override
-  String getTitle() => 'Paragens';
+  String getTitle() => 'Autocarros';
 
   @override
-  onClick(BuildContext context) => Navigator.pushNamed(context, '/Paragens');
+  onClick(BuildContext context) => Navigator.pushNamed(context, '/' + Constants.navStops);
 
   @override
   Widget buildCardContent(BuildContext context) {
@@ -53,7 +54,7 @@ class BusStopCard extends GenericCard {
             child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: <Widget>[
-                  Text('Configura as tuas paragens',
+                  Text('Configura os teus autocarros',
                       maxLines: 1,
                       overflow: TextOverflow.ellipsis,
                       style: Theme.of(context)