+++ changelog.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- User faculty storage in shared preferences
 - Send bug reports and app crashes details to Sentry
 - Clickable checkbox rows
+- Fetch restaurants from sigarra and stored locally
+- Dropdown to choose the student's faculty in the login page
+- Calendar button to add an event to the calendar with the details of a exam
+- Drag icon appears on widgets during editing mode
 
 ### Fixed
 
@@ -22,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated Android's `targetSdkVersion` to 30
 - Changed the Checkbox background to dark red
+- Changed 'Paragens' to 'Autocarros'
 
 ## [1.1.0] - 2021-04-18